test: created __mocks__ folder to store all mock
