+++ web/src/test/__mocks__/canvas.ts
@@ -1,0 +1,5 @@
+import { vi } from "vitest";
+
+vi.stubGlobal("jest", vi);
+
+import "jest-canvas-mock";
