+++ web/src/test/__mocks__/global.ts
@@ -1,0 +1,27 @@
+import { vi } from "vitest";
+
+vi.stubGlobal("open", vi.fn());
+
+vi.stubGlobal("scrollTo", vi.fn());
+
+vi.stubGlobal(
+  "matchMedia",
+  vi.fn().mockImplementation((query) => ({
+    matches: false,
+    media: query,
+    onchange: null,
+    addListener: vi.fn(), // Deprecated
+    removeListener: vi.fn(), // Deprecated
+    addEventListener: vi.fn(),
+    removeEventListener: vi.fn(),
+    dispatchEvent: vi.fn(),
+  }))
+);
+
+class ResizeObserver {
+  observe() {}
+  unobserve() {}
+  disconnect() {}
+}
+
+vi.stubGlobal("ResizeObserver", ResizeObserver);
