+++ web/src/test/__mocks__/index.ts
@@ -1,0 +1,5 @@
+import "./canvas";
+import "./plotly";
+import "./global";
+import "./rudder-sdk";
+import "./open-replay-tracker";
