+++ web/src/test/__mocks__/open-replay-tracker.ts
@@ -1,0 +1,8 @@
+import { vi } from "vitest";
+
+class Tracker {
+  start() {}
+  setUserID() {}
+}
+
+vi.mock("@openreplay/tracker", () => Tracker);
