+++ web/src/test/__mocks__/plotly.ts
@@ -1,0 +1,14 @@
+import { vi } from "vitest";
+
+vi.mock("plotly.js", () => {
+  return {
+    default: {
+      newPlot: vi.fn((ref) => {
+        ref.on = vi.fn();
+      }),
+      addTraces: vi.fn(),
+      restyle: vi.fn(),
+      relayout: vi.fn(),
+    },
+  };
+});
