+++ web/src/test/__mocks__/rudder-sdk.ts
@@ -1,0 +1,9 @@
+import { vi } from "vitest";
+
+vi.mock("rudder-sdk-js", () => {
+  return {
+    ready: vi.fn(),
+    load: vi.fn(),
+    track: vi.fn(),
+  };
+});
